Cambios Registro Estudiante con repest

+++ HenriPittier/src/comun/Rep_Est.java
@@ -10,7 +10,7 @@ public class Rep_Est {
     private int cedula;
     private String primerNombre_Rep;
     private String primerApellido_Rep;
-    private int cedulaEscolar;
+    private long cedulaEscolar;
     private String primerNombre_Est;
     private String primerApellido_Est;
     private int error;
@@ -23,14 +23,14 @@ public class Rep_Est {
         this.cedula = cedula;
     }
 
-    public Rep_Est(int cedula, int cedulaEscolar) {
+    public Rep_Est(int cedula, long cedulaEscolar) {
         this.cedula = cedula;
         this.cedulaEscolar = cedulaEscolar;
     }
     
     
 
-    public Rep_Est(int cedula, String primerNombre_Rep, String primerApellido_Rep, int cedulaEscolar, String primerNombre_Est, String primerApellido_est) {
+    public Rep_Est(int cedula, String primerNombre_Rep, String primerApellido_Rep, long cedulaEscolar, String primerNombre_Est, String primerApellido_est) {
         this.cedula = cedula;
         this.primerNombre_Rep = primerNombre_Rep;
         this.primerApellido_Rep = primerApellido_Rep;
@@ -63,11 +63,11 @@ public class Rep_Est {
         this.primerApellido_Rep = primerApellido_Rep;
     }
 
-    public int getCedulaEscolar() {
+    public long getCedulaEscolar() {
         return cedulaEscolar;
     }
 
-    public void setCedulaEscolar(int cedulaEscolar) {
+    public void setCedulaEscolar(long cedulaEscolar) {
         this.cedulaEscolar = cedulaEscolar;
     }
